fix: PHPStan level 7 — type safety (96 → 54)
- Add fopen() false checks in ExportController, TaxExport, ProductCsvService,
  BackupDatabases, SurveyResults
- Cast QrCode color args and imagecolorallocate args to int
- Cast float|int|string to float for property assignments (FinanceSummary)
- Fix ucfirst on enum types (use ->value)
- Cast mktime return to int, filesize to int
- Handle glob() returning false (?: [])
- Cast ob_get_clean() and hash_hmac payload to string
- Fix number_format on mixed numeric types
- Bump PHPStan to level 7

+++ app/Filament/Pages/FinanceSummary.php
@@ -128,7 +128,7 @@ class FinanceSummary extends Page
 
             $this->monthlyBreakdown->push([
                 'month' => $month,
-                'month_name' => date('F', mktime(0, 0, 0, $month, 1)),
+                'month_name' => date('F', (int) mktime(0, 0, 0, $month, 1)),
                 'revenue' => $totalMonthRevenue,
                 'expenses' => $monthExpenses,
                 'net' => $totalMonthRevenue - $monthExpenses,
